pyproject.toml
[project]
name = "brilliance-admin"
version = "0.44.7"
description = 'Simple and lightweight data managment framework powered by FastAPI and Vue3 Vuetify all-in-one. Some call it heavenly in its brilliance.'
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "asgiref>=3.11",
    "fastapi>=0.115",
    "jinja2>=3.1",
    "PyYAML>=6.0",
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["brilliance_admin*"]

[tool.setuptools.package-data]
brilliance_admin = [
    "static/**/*",
    "templates/**/*",
    "locales/**/*.yml",
]

[project.optional-dependencies]
example = [
    "uvicorn>=0.34.0",
    "faker>=38.2.0",
    "pyjwt>=2.10.1",
    "structlog>=25.5.0",
    "rich>=14.2.0",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.12.0",
    "twine",
]
tests = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.28.1",
    "pytest-mock>=3.15.1",
    "sqlalchemy>=2.0.41",
    "aiosqlite>=0.22.1",
    "factory-boy>=3.3.3",
    "pyjwt>=2.10.1",
]
scalar = [
    "scalar-fastapi>=1.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
